atualização do polling

+++ arp_robot/arp_base/include/arp_base/roboteq_driver/channel.h
@@ -113,7 +113,7 @@ protected:
 
   FeedBack msg_;
 
-  int polling_timeout_;
+  double polling_timeout_;
 };
 
 enum ModePosition
